Add a base template for non admin views
This commit also introduces Twitter bootstrap and django-compressor,
the principal dependencies for doing frontend on Zenon.

The SASS version of Twitter bootstrap is used. Any customization
to the default Bootstrap templates should be done in a separate
scss file to avoid tampering with the framework files and breaking
any future upgrades.

Bootstrap is loaded with django-compressor in the base template.
django-compressor will compile the scss files, minify and combine
them.

ref #164, #179