Allow exceptions where we load jQuery in <head>
In the issue submission form, we have inline JS being injected by a
widget. That injected JS requires jQuery, and it fails because jQuery is
only loaded at the bottom of the page.

This commit adds a new template context option, `put_js_in_head` which
moves JS includes in the head rather than at the bottom of the body.