article: replace title property by ArticleTitle FK
An article may have a principal titles and parallel titles.
To support this, replace the title property by a 1-to-M
relationship between Article and ArticleTitle. ArticleTitle
defines the title, if the title is parallel or principal, and
the language of the title.

Provide three migrations to migrate the existing Article
objects to the new format. The first add the FK between
Article and ArticleTitle. The seconds converts the title
attributes to ArticleTitle objects. The third deletes the
title property on the Article object.

Modify the import script to create ArticleTitle objects.

To ensure backward compatibility, provide a cached title property
that retrieves the first non-parallel title of the article.

ref #1948