issue: add is_published
Because an issue may be in Fedora without being published, add
is_published to explicitely determine if an issue is published or not.
The ``date_published`` field of an issue is no longer used to determine
if an issue is published. The default value of is_published is True in
model factories, since issues will be published in the majority of
cases.

ref rm#2522