solrq: implement filter_query support
There are two ways to perform a Solr query: with ``q`` or with ``fq``.
``q`` is the standard way to perform a query, while ``fq`` is used to
restrict the super set without influencing the score.

The ``filter_query`` methods has the same behaviour as ``filter``. It
will build the querystring and return a new ``Query`` object.

for more info see https://wiki.apache.org/solr/CommonQueryParameters