support multiple individual subscriptions
Implement a UserSubscriptions model that serves the purpose of
aggregating all the user's subscriptions and finding the appropriate one
for a given Article, Issue or Journal object.

UserSubscriptions is set directly on the request object by the
SubscriptionMiddleware.

The subscriptions are implemented as a list and not as a set, meaning
that the order in which they are added is important. If two different
subscriptions provide access to the same content object, only the first
one to be added will be used.

ref #1529