stats: create new, common and legacy packages
``new`` contains the new implementation that uses influxdb. ``legacy``
contains views that calls to the legacy implementation of the stats
services. ``common`` includes any content that needs to be shared
between the new and legacy implementation (e.g. landing page view)

use a feature switch to dispatch to the ``new`` or ``legacy``
implementation.

ref #1511