fix infinite loop in tests
The TestCases mock return values from Fedora. django-model-translation
patches all querysets on registered models. When a query is performed
on a patched model, django-model-translation parses the query from
left to right, using lhs.next. Because models are mocked with no
specified return value, this results as an infinite loop, because
``lhs.next`` will never return None (and always a mock)

This commit fixes the infinite loop issue by making sure that every
mocked Issue object in the tests return a specified value for attributes
that will be used in a query.

ref #1599