Improve restrictions import process
We now properly delete and update subscriptions from upstream. Because
Victor is our *only* source of subscription information, the simplest
approach was to delete all and repopulate on each import.

The granularity of the import changed a bit. Instead of iterating mainly
through subscriptions, we iterate through subscribers, and then through
subscriptions. This allows us to atomically delete subscriptions per
subscriber. Because we wrap all of this in a transaction, at no moment
during the import do we end up with an empty subscription list that
could disturb production operations.

fixes #1459