Make requirements.txt be managed by pip-tools
Move the old `requirements.txt` to `requirements.in` and have
`requirements.txt` be generated by `pip-compile`, a part of the
`pip-tools` package.

The previous setup for dependencies management was very shaky and mixed
transitive dependencies with real ones. Some transitive packages were
installed to the correct (or incorrect) version by *chance*, only
because of the install order.

`pip-compile` brings order into all this and will allow us to update
dependencies safely, transparently and, most importantly, *frequently*!

This commit is proposed as an alternative to the `add-pipfile` branch
mostly for its "minimal change" approach. With the approach proposed
here, nothing changes deployment-wise. We still deploy with `pip install
-r requirements.txt` with no additional dependency. Moreover, I have
reservations about the complexity of the `pipenv` tool (to be
discussed).

I have been very conservative with version pinning in `requirements.in`.
My approach was that if the package's current version was more than a
minor version away from the old pinning, I would add an upper bound.

Generally, except for the `django` dependency, we will want to avoid
these upper bounds to ensure up-to-date dependencies, but because we
have so many out-of-date dependencies, it seemed dangerous to update so
many dependencies at once. I expect that, over the time, we'll remove
these upper bounds one by one after properly ensuring the stability of
the system.

The only exception is the `weasyprint` dependency which, because of a
transitive dependency conflict, had to be updated from 0.26 to 0.47 in
order for automated tests to pass. Before deploying this, we'll need to
tests `weasyprint`-enabled features more particularly.