De-templatetag-ize render_article
Move the `render_article` template tag into `ArticleDetailView` and
`ArticleSummaryView`. We do this to avoid having our errors silently
ignored by Django (that's what it does when an exception is raised in a
templatetag).

Article XSL rendering is too big and complex to keep it out of our
Sentry debugging capabilities.

Other than the move itself, I've tried to avoid changing anything in the
code itself. It mostly stayed as-is. I had to modify the tests so that
it instantiate and mock a view.

Also, update Raven. Our previous raven version triggered an error when
trying to handle exceptions.

ref https://gitlab.erudit.org/erudit/support/issues/163