oai imports: force sickle encoding to utf-8
We've been running on a private fork of sickle forcing encoding to utf-8
for a long while now but sickle 0.6 now allows the same encoding forcing
without the need for a private fork.

This commit bumps the sickle requirement and use the new feature. We
will then be able to remove the private fork requirement from eruditorg.