fix raw PDF tests
Do not use a journal that is in the ``erudit`` collection. Since
the changes to ``published_issues``, we need to mock the return value of
:meth:`get_published_issues_pid` to calculate the embargo period of the
journal.

Because managing the embargo period is unnecessary for those tests,
using a journal that is in another collection is simpler than mocking
the embargo and subscription objects.

ref #1310