journal detail: link to external issue url if set
When `external_url` is set on an `Issue`, use this as our link when
rendering the journal detail view.

As a side note: I don't like the way functional tests are made by
accessing context variables. This is not functional testing because
is accesses a private API that is subject to change.

ref #1651