Use erudit-core as a submodule
It's *much* easier to work with erudit-core and eruditorg in parallel
when working with submodule, especially when pushing WIP branches
that also involve WIP branches in the core.

The principle is rather simple:

1. Add erudit-core submodule
2. Remove erudit-core from pip deps
3. Add a `erudit` symlink to `erudit-core/erudit` in `eruditorg`
4. Add a tox env to run tests in erudit-core
5. Include erudit core dependencies in `requirements.in`

It works well so far. That makes setup/tox-related stuff in erudit-core
obsolete but we're in no rush to remove them.