Fix bug where import restriction transactions weren't commiting
Raising an exception to escape the inner function when a journal
couldn't be found wasn't a good idea because it caused
`transaction.atomic` to rollback an otherwise good transaction. Now, we
return from the function instead.

Also, made the subscription wiping process a bit more precise to avoid
deleting stuff that doesn't need to, that is, the
`JournalAccessSubscription` instance.

fixes #1459