tests: Deparametrize test_can_embed_the_latest_issue_in_the_context
It was a sloppy idea to parametrize a test that didn't have exactly the
same asserts. I wanted to avoid setup code copy/paste, but didn't do the
right thing.

The right thing was to properly take advantage of our factories to
reduce the size of the setup code. Then, we can happily copy/paste.

Also, deduplicate journal reverse url acquisition code.