Remove django-filters dependency
We don't use it anymore. It was introduced in
56c31d64cbeb9323d3ed12d09654b241f2366ef2 but there's now no reference to
`django_filters` in the code now. Removing.

We had `DjangoFilterBackend` as a default backend but after having read
http://www.django-rest-framework.org/api-guide/filtering/#djangofilterbackend
I came to the conclusion that it wasn't needed because our only
`ListAPIView`, `EruditDocumentListAPIView` doesn't have any
`filter_fields` attribute or doesn't directly import `django_filters`
for "Advanced Usage".

Also, the search seems to work as it was before. My guess is that if
something was gonna break, it would break obviously.