search: instantiate "real_object" after deserialization
The end goal being to get rid of django-rest-framework.

The first step is this commit where we get rid of all serializers
except one, `GenericSolrDocumentSerializer` (which simply serialize the
stuff it gets of solr).

The work that was previously done by those removed serializers, that is,
augment solr data with data coming from our django database, is now done
later, after the deserialization (`json_load`) phase. At that stage, we
run a django query to get our corresponding model (for articles and
thesis).

We *could* simply use those models and set `real_object` with it but
that would imply prfound changes in the complex search result templates.
Because this is a big refactoring, we go one step at a time so I
re-implemented old serializer's `real_object` API in temporary shims so
that I don't have to touch the templates, for now.