Reduce usage of Article.authors to a bare minimum
Stop using `Article.authors` everywhere but at one place: the authors
index view. Ultimately, we want to "de-django-ify" our article and issue
models so that we rely on first-hand data sources (fedora, solr)
directly instead of syncing that info regularly. This is done step by
step.

In this step, it was the `authors` field. I hoped to remove it entirely,
but I stumbled upon the authors index which at this time can't be
realistically implemented in any way other than the `Author` django
model. So, with this commit, that's now the only place where we use it.

The most exciting benefit of this change is on the search side: we still
use `Article` to render authors listings in the search results, but
*only* for articles in Fedora. When rendering results for articles
outside of fedora, we use authors data we get from solr.

This means that when we're finished going through all `Article` fields
like we did with `Author`, we won't need to sync `Article` outside of
fedora again because we will not be using any of that information
anymore.

This commit introduces new test mechanisms. With the de-django-ification
of `Article` comes a new challenge: the generation of test articles with
author information. Until now, we used the `Author` model for that. This
commits adds a new way to create test articles with author information:
`FakeFedoraAPI`'s `tweak_article()`. This function returns a "tweaker"
class that starts with the default article fixture DOM and adds helper
function to manipulate that DOM and replace author information with
something to your liking.

This method of creating test data will probably gain importance as we
de-django-ify `Article` and `Issue` further.

Last thing: in search, add a `can_cite()` method to search proxy models.
Sorry this is bundled in that commit, but that bug fix for #1491 was
needed because what was previously causing simple 404s now cause the
search results to stop rendering.