Generate author index through solr queries
With this commit, we stop using `Author`'s relationship with `Article`
to generate author index for a journal. Instead, we use information from
our solr instance.

This has the potential to remove a lot of dead code, but I haven't
removed it yet so that we can compare our two indexing methods with a
temporary `benchtest_article_indexes` script.

The script show a lot of minor differences between indexes, but this is
mostly dues to a differing inaccuracies set between the solr method and
the fedora-sync one (both method have inaccuracies).

These results we have now are much closer to what we had on Retro (which
also generates author indexes from solr).

This commit also introduces a brand new "fake API" for solr, similar to
what was done for fedora. This allows us to fake results at a much lower
level than what we would normally do through naive mocking and thus
expand our coverage for author index tests. This also lays a foundation
to expand coverage for all our solr-related code.

This commit also changes the context for a potential fix in
erudit/support#197