Update to django 1.11
Not many adjustments in eruditorg itself.

There's `infoimg` that started to be queried during the tests for some
reason and we want to check that it exists before fetching its content
(that would have been a good idea even without the django update).

Then, there's warnings that began popping up about undefined orders in
object listing. Those are nice warnings. Because we had an undefined
order before, I didn't ask myself too many questions and went with `pk`
ordering. It might not be a very good ordering, but at least it's
explicit now.