Use Solr ids as a basis for saved citations lists
Instead of storing `EruditDocument` pks in saved citations lists, we now
store Solr ids.

We need to do this because of our recent de-django-isation of `Article`:
external articles saved in citations list became broken.

This is a monster commit, sorry, I didn't expect it to grow so large
when I began. The size of this commit is mainly dues to two factors:

1. Test harnessing. Tests related to saved citations had to be properly
solr-aware. I had to go fetch my Solr fake API code from the
yet-to-be-merged `solr-author-index` branch. It's a big chunk.

2. Frontend changes. Frontend code assumes simple numerical ids. With
Solr ids, that assumption breaks. I had to modify urls, views, JS to
adapt to that.

This commit takes care of the migration of both types of citation lists:
in-DB and in-session. For DB-stored lists, it's easy: it's a data
migration. For sessions, I had to add code to handle legacy sessions.