journal: cache get_titles
cache the output of the get_titles method of Journal. This method is
costly:

In our fedora model, the current journal title is not set at the Series
level, but at the Publications level. In order to retrieve the current
journal title, we need to:

1. Retrieve the Series object in Fedora
2. Retrieve its PUBLICATIONS datastream and find the last issue
3. Retrieve the Publication object
4. Retrieve its SUMMARY datastream

This is four hits to Fedora to get one title. Fedora objects are cached
in the database, but I believe it makes sense to cache some method
outputs.

ref #1334