fake solr: use luqum to parse solr queries
The regex approach had limit that we hit really fast. luqum is a parser
that gives us an AST with is much easier to normalize and pattern-match.

Sure, it's a bit more complex but I think it's worth it.

If we ever hit a complexity wall with this approach too, we should
replace the fake solr with a real instance.