Sort journals using the "locale" module
We have problems with sorting journal names with accents and brackets
and articles (la le les...). This commit introduces a fix for these,
mostly piggy-backing on the `locale` module. Collation rules in the
`fr_CA` locale already mostly does what we want. All that is left to do
after that is to handle article prefixes, which we do on our own.

ref #1006