Remove most references to Article.objects
In preparation for a de-django-ification of the whole `Article` model, I
stripped most of the references to `Article.objects`, replacing it with
direct queries to Fedora and Solr.

Most of the work here involved adjusting test support code without
having to rewrite too many of them. Now, each time we create an article
through `ArticleFactory`, an entry is automatically created in our fake
Solr and a reference to the article is automatically added to the fake
publication in fedora (so that when we query for a list of articles for
a given issue in our new "live fedora" code, we get that list properly)