De-django-ify Article
`Article` is not a django model anymore. It's now a proxy to its
`erudit_object` that keeps an API that is similar to what the old django
model offered.

An `Article` now always refers to something that exists in fedora and
there is no concept of `external_url` anymore. It used to be useful in
some search results, but it's not the case anymore. All that logic has been
stripped.

The citations transition code has also been removed because it required
access to old article PKs which we don't have anymore.

This is the end of `EruditDocument` and polymorphic models. With
`Thesis` recently gone, `Article` was the last subclass.