Make restriction API permanent
It seems that the idea of adding a temporary legacy API for restrictions
wasn't so good. It's not even deployed yet and we're already modifying
consumers to use more precise restriction dates.

So, to keep things easy, I've made the API permanent and added
`embargo_date` to the `revue` element.