Remove SolrServiceRequiredMixin and FedoraServiceRequiredMixin
The main drive for this change was to remove the unnecessary
`SOLR_ADMIN` config which was needed for `SolrServiceRequiredMixin`.

`SolrServiceRequiredMixin` was being required only for one internal
reporting view and this machinery seemed deeply useless to me.

And then I saw `FedoraServiceRequiredMixin` and thought that I could
remove that too. After all, if either Solr or Fedora are down, eruditorg
is down anyways. Those checks are unnecessary.