import: do not assume non unique codes
Import was first built on the assumption that codes are not unique.
The import script tries to load journal by localidentifier, and if that
doesn't work, a new journal is created. This journal is then assigned a
code. If the code already exists, it is incremented by 1.

This doesn't work so well with the journal creation process because
journals are created in the admin before we know the localidentifier.
This means that whenever we create a new journal manually, the import
script will create a copy with its code incremented by 1.

This commit fixes that by having the import script:

1. Assume that codes are unique
2. First try to retrieve a journal by localidentifier
3. If the journal doesn't exist, try retrieving it by code
4. Always set the localidentifier to whatever is returned by fedora

This commit is tested by registering datastreams in the FakeAPI.
Because the FakeAPI needs to be monkeypatched everywhere it is
imported in production code, this commit monkeypatches it in the
import_journals_from_fedora commands (see conftest.py.)

Ref: #1964