Log warnings when Issue is not in Fedora
Now that we have dropped OAI support, we can safely expect that all of
our issues will be in Fedora and that we can use ``erudit_object`` to
retrieve their information. However, before we drop support for issues
that are not in Fedora, deprecate the methods that access the database
fields and log warnings if we use them.

Ref #1994