Remove provided_by_fedora property
Remove deprecated ``provided_by_fedora`` property and replace all calls
with ``is_in_fedora`` calls.

This commit also changes the logic of ``provided_by_fedora``,
``get_full_identifier`` and ``is_in_fedora``.

The logic of the methods was the following:

* ``provided_by_fedora``: ``True``  if the ``Journal`` object has no external url,
has a ``localidentifier`` and is part of a ``Collection`` that has a
``localidentifier``

* ``is_in_fedora`` if the ``Journal`` has a full identifier and if that
full identifier can be used to fetch an object in the Fedora repository.

* ``get_full_identifier`` returns a full identifier if the model has no
external url and if it and his parents (e.g. in the case of ``Issue``:
``Journal`` and ``Collection`` all have localidentifiers.)

This approach has multiple problems:

1. ``get_full_identifier`` should not depend on the object having an external url
or not. If an object has a localidentifier, is part of a hierarchy of objects that
all have localidentifiers, it should always have a full identifier, and
``is_in_fedora`` should always return ``True`` regardless of the
external url. Setting an external url on an object does not remove it
from Fedora.

2. ``provided_by_fedora`` and ``is_in_fedora`` were both used to check
if the object is in fedora and were not used with much distinction.

This commits ensures that:

* ``get_full_identifier`` returns the full identifier regardless of the
presence of an external url

* ``is_in_fedora`` always return ``True`` if the article has a full
identifier and we can retrieve it.

ref #1994