Implement canonical urls as a middleware
Implement PolyglotLocaleMiddleware, a LocaleManager that speaks all the
languages!

Django's LocaleMiddleware handling of non language prefixed urls can be
problematic when the default language is also prefixed, as it only tries
to find the pattern in the user's language urlconf.

This commit modifies Django's LocaleMiddleware to check for a non
language prefix url match in *all* the urlconf. If it finds one, it will
then translate it to the user's language and redirect there.

For example:

- If the user's language is 'fr'
- And the following urls are equivalent:
        'fr' -> /fr/revues
        'en' -> /en/journals

- With LocaleMiddleware, a request for /journals would redirect the user
to /fr/journals and ultimately return the 404 page.

- With PolyglotLocaleMiddleware, a request for /journals will find that
/journals exists in the 'en' urlconf, translate '/en/journals' to
'/fr/revues' and redirect there.

ref #1934