Use Django's LoginRequiredMixin
Django 1.9 introduced LoginRequiredMixin, so there's no need to roll
our own anymore. This commit replace our LoginRequiredMixin with
Django's.

Using LoginRequiredMixin broke tests, because it uses ``has_permission``
to determine if the user has access to a view and return a 302 or 403
depending on the value of ``self.raise_exception``. In some cases, the
code was relying on our ``LoginRequiredMixin`` to return a 302 and
``has_permission`` to return a 403.

This has been modified so that ``has_permission`` will explicitely
return a 302.

Ref: #2122