Remove hint CSS on footnotes' links
The Hint.css library is used on footnotes' links to display the begining of the footnote on mouse
hover. The library cause some display problems where quotes or periods appear on new lines and where
non-breaking spaces are ignored. Let's replace the library by the siple title tag that does almost
the same thing.

ref #2010, #2105, #2116, #2138