Don't encode URLs and fix smooth_scroll() JS function instead.
The problem with non-working links was not that they were not encoded properly, since they worked
on retro without manual encoding, but because of our JS for smooth scrolling when clicking on same
page anchors (footnotes). This function broke any links with an anchor part (#) that was not in the
same page.

ref #2057