Do not cache the back issue link with the journal info template fragment
Caching the back issue link with the journal info template fragment causes problems for journals
with no published issues. When they publish their first issue, the back issue link does not appear
because it is cached with the journal info template fragment which is not invalidated unless it is
modified.

ref #2329