fix coverpage encoding errors
There was a possibility of uncaught encoding errors because reportlab
was trying to encode its strings using pikepdf's pdfdoc codec instead of
its own.

The python codecs lookup function iterates through a list of search
function to find the appropate codec for an encoding.  pikepdf and
reportlab both try to register their version of the pdfdoc codec.
Whoever registers it first will set it for both libraries.

In our case, pikepdf gets to register its plugin first because it is
imported first.

Instead of changing the import order by importing reportlab early in the
request cycle, we decide to sanitize the input before generating the
coverapge.

This commit does the following:

* Log any conversion error as a warning
* Strip the problem characters from the string

Refs:

- https://docs.python.org/3/library/codecs.html