Fix errors when dealing with HTTP OPTIONS requests
The cause of this bug is ``ContentAccessCheckMixin.get_content``
returning ``None`` instead of returning the target content when
dealing with HTTP OPTIONS request because its ``object`` attribute
has not been initialized at this moment. This results to a
``provides_access_to`` with both article, issue and journal set to
``None``. The ``object`` attribute is initialized by the ``get_object``
method on its first call.

The reason why ``self.object`` is initialized when dealing with
GET requests is because ``PrepublicationTokenRequiredMixin.get`` calls
``get_object`` and initializes the value, while
``PrepublicationTokenRequiredMixin.get`` is not called for HTTP OPTIONS
requests.

This commit fixes the issue by calling ``get_object`` and not returning
the object. It does not fix the underlying issue of having a getter
method that has side effects.