Convert ip addresss to int
Use the ipaddress module to convert ip addresses to their integer
representation. Use the integer representation of the ip address to
determine if a user has subscriptions. This lets us index the
ip_start_int and ip_end_int columns and optimise the query.

We do not want the users to input or to even see the integer
representations of the ip addresses. This field is populated when save()
is called on the model.

ref: #2385, #2392