Always define main_title in JournalDetailView
Because the main_title context key is not always defined in
JournalDetailView, we need to test it before using it in the
template code. If it's present, we use it, otherwise we use
the journal name stored in the database.

This commit ensures that main_title is always defined in
JournalDetailView's context so we don't have to test it in the template
code. It also simplify the journal_detail template by always using
main_title

ref support#478