Provide a Dockerfile and docker-compose files
Add a development and a production Dockerfile to the project, with a
docker-compose file to run the project in development.

The development dockerfile:

- starts gunicorn in automatic reload mode
- it sets std_open and tty to true, to allow for ipdb.set_trace()
- it configures http_proxy to use a socks proxy on the host on port
  8181. This means that the compose file assumes that you have a proxy
  running on the host.
- it uses the development django.env environment variables (with
  DEBUG=True for example)

No production django.env is provided, because we do not want to commit
the production settings in this repository.

If use the compose file for your development environment, you might want
override some of the settings with your own. For this purpose, a local.yml.sample
file is provided. Copy it to local.yml (it is in gitignore) and start it
like this:

docker-compose -f compose/dev.yml -f compose/local.yml up

warning: the production Dockerfile is not fully ready for production --
we still need to add some font packages and for edge cases.

ref: #2539