Add issue_coverpage_url templatetag
The issue_coverpage_url templatetag takes an issue as an argument and
will return a cacheable url for its coverpage. The url includes the
timestamp of the last modification of the coverpage in order to expire
the cache whenever the coverpage changes. These urls can be cached
permanently.

These coverpage are meant to be served from a CDN. However, we might now
want to use this in development environments, so the
FEDORA_ASSETS_EXTERNAL_URL is configurable. If it is set to None, the
templatetag will return a relative url will be returned; if it is
defined, an absolute url will be returned.

This url has to be set directly in base.py because we do not want this
to be a i18n_url.

Finally, this commit adds two new targets in the fake Fedora API to mock
coverpage datastream content and history.

ref: #2549