FedoraMixin: add get_erudit_content_url
Let models return the URL of their content. This URL can then be used by
get_erudit_object.

The eulfedora implementation checks for the datastream existences before
issuing the request, resulting in unnecessary http calls to Fedora.

To support this change, we need to change the way we mock articles in
the test. We do so by:

- monkeypatching requests.get to go through the Fake Fedora API
- mocking article.erudit_object directly when we need an alternative
  erudit_object, instead of mocking deeply in the Fedora implementation.

ref: #2679