metrics: add article_views_by_subscription_type
Add a metric to know how many article views we have, by subscription
type (individual / institutional)

There is a known issue with this metric: if the article has a PDF
viewer, it will be counted as two views: one for the html version, e.g.
123567ar, and one for the PDF version of the article. I don't think we
have a way of making this distinction at the moment, and we have the
same issue with logging HTTP requests in the subscription middleware.

ref: #2709