subscription: remove Organisation fields
These fields were added early in the project, when we had the ambition
of using the portal as a small CRM. The idea was to have two models:
Organisation, where we sould store all the client information, and
LegacyOrganisationProfile for compatibility with the legacy databases.

Practice showed that this idea was not good: we never used the portal as
a small CRM, so these fields have never been used. Moreover, the
information contained in the legacy database is not "legacy" after all,
and still used widely.

This commit is the first step towards merging Organisation and
LegacyOrganisationProfile by removing the spurious Organisation fields
and adding the LegacyOrganisationProfile fields directly in the
Organisation model.

ref: #2712