get_cached_datastream_content(): Remove its use from get_articles_from_fedora()
We do not need to use get_cached_datastream_content() to get the issue's SUMMARY since the
erudit_object already have access to it and it also caches it. We also do not need to parse the
SUMMARY xml since the erudit_object already does it for us with get_summary_articles().

red #2711