Use get_cached_datastream_content() to get the articles' PDFs instead of eulfedora
Instead of using eulfedora to check if a PDF exists and thus making a request to Fedora only for
that, get it directly from its URL and cache it. If there is no PDF, the content will be None and
we can easily test for that to see if a PDF exists.

ref #2711