PolyglotLocaleMiddleware: set Content-Length
CommonMiddleware sets the Content-Length of the response.
PolyglotLocaleMiddleware may create a HTTP redirect response and pass it
through the rest of the middleware chain.

If this is the case, we need to explicitely set the Content-Length of
the response to 0. Otherwise, it may be rejected downstream by a proxy
that strictly adheres to RFC2616, like traefik.

If the Content-Length of a HTTP 301/302 response is not 0, a body is
expected. Because our redirect responses do not set bodies, we need to
set the Content-Length to 0.