subscription middleware: use left-most proxy
The middleware was modified in erudit/support#205 to use the right-most
proxy. This modification was made to support a case where a client's
proxy had multiple layers, like this:

    external    +  internal
                |
                +
    +--------+     +-----+     +---+
    |P1|..|Pn| +-> |httpd| +-> |app|
    +--------+     +-----+     +---+

Where: external is the external network, internal is Érudit's network,
and P1 to Pn is Proxy 1 to Proxy N. In that case, the HTTP request
coming from Pn would have an X-Forwarded-For header containg the proxy
chain as per RFC 7239. If we use the left-most proxy, we would get the
IP forwarded by P1, which not the IP address of the proxy server, and
deny the subscription.

Now if we have a setup where we have our own proxy in the stack:

    external    +  internal
                |
                +
    +--------+     +-----+     +--------+     +---+
    |P1|..|Pn| +-> |httpd| +-> |P1|..|Pn| +-> |app|
    +--------+     +-----+     +--------+     +---+

If we use the right-most proxy of the X-Forwarded-For header, this will
give us the address of Pn-1 on the internal side. Obviously, this is not
what we want. What we want is the address of Pn on the external side. To
achieve this, we need to:

1) Have the httpd server strip the X-Forwarded-For header given to us by
   the client.  We have no reason to care about the proxy chain on the
   client side, and we have no reason to trust this information.

This will leave us with an X-Forwarded-For chain that contains only our
internal proxies. We can then do:

2) Revert to using the IP address of the left-most proxy. That means
   using the X-Forwarded-For header set by P1, which should be the
   address of Pn on the external side.

Ref: erudit/support#205
Ref: erudit/ti/infra/infra#631
Ref: https://tools.ietf.org/html/rfc7239#section-7.5