Fix how we check if an article has a PDF
Instead of downloading the full PDF just to see if it exists, get the list of datastream and check
if there's a PDF in that list.

To do so, we add the new function `has_datastream` which takes the datastream's name and it makes a
request to Fedora to get the list of datastream for this object and check if the datasteam name is
in the list.

ref #2708