coverpage: force translation of article title
3fdd3c23 changed the return type of Article.get_title from
Optional[Union[Promise, str]] to Union[Promise, str], where Promise
is ugettext_lazy's Promise of the final str.

The intent is to let Article.get_title return the translated
"[Article sans titre]" string when the title is None.

However some callers of this function, like reportlab's PDFString,
will work when given None and raise an exception when given a promise.

This commit fixes the coverpage case by forcing the conversion to str
and sending the result to reportlab.

fix: #2759