Remove django-reversion-compare and use diff-match-patch instead
The package django-reversion-compare is blocking us to update Django to the latest version because
it has dropped support for Python 3.6 while Django has not. We were using this package only to send
emails with the diffs between version of JournalInfo. We can simply use the package diff-match-patch
to generate the diffs ourselves.

ref #2796