Add send failed password reset email to non registered users
Add a new password reset procedure to send the user an email informing
that the password reset has failed if the entered email does not
belong to the list of registered emails. If the email belongs to the
list of registered emails the procedure remains as it was, being,
sending an email with a link to reset the password.

Ref.: #2746