Add CSS rule and JS to allow scrolling filter sidebar in low screens
For screens that are to low (heights smaller than 656px) the filter sidebar
menu is not visible in its totality. If the user selects a lot of itens in
the `By discipline` field, it encreases the total height of the filter so
the problem of visibility can also happen for higher screens (greater than
655px). To fix the problem we added an CSS rule and a JS function to allow
scrolling through the menu.

Ref.: #2807